Update pod resource to support containers
Update podman::pod resource to support deploying containers as part of
a pod. Containers are defined as a hash of containers that is used to
declare individual container resources. Systemd service units will be
created to automatically start the pod and its containers. Any changes
to the pod flags or containers hashes will trigger the entire pod to be
redeployed.

The podman::container resource was updated include a private parameter
called pod for declaring containers that are part of a pod. The
podman::container resource will not create systemd service units if the
pod parameter is defined. The systemd service units for the container
will be created by the podman::pod resource. The pod parameter is not
intended to be used on its own.